Añadiendo evento a bd y cargandolos en recycler

+++ EventosCompartidos/app/src/main/java/model/EventoGestion.kt
@@ -4,5 +4,5 @@ data class EventoGestion(
     var nombre: String,
     var fecha: String,
     var hora: String,
-    var numAsistentes: Int
+    var numAsistentes: Long
 )
